On startup, attempt to find the offset of the binary
this should set r12 correctly so that the TOC can be properly loaded.
For non-pie binaries, the offset is 0.
For pie binaries, it attempts to compute the off offset of the link
address and the execution address, this value should be loaded
into r12 as an initial base. Could this be loaded from dl_data instead?

Update:
Once asm code blocks are updated to have proper .uselocal entry, this
will not need to be filled in for early calls.
static
  pie: dl_data may be useable ?
  non-pie: r12 value is 0x0
for dynamic binaries either ld.so could fill
  pie: either ld.so could fill in r12 with the offset or use this trick
  non-pie: r12 value is 0x0.