Enable interrupts on riscv before calling WFI.
On ARM64, WFI will sleep until there is a pending interrupt even if the
external interrupt is disabled. It will wake up and allow the system
to execute some code (eg to come out of low power modes) before the
interrupt is enabled allowing the interrupt to be serviced.
It appears that the system will just hang with pending external and timer
interrupts if wfi is called with external interrupts are disabled.